[refactor] : 모임 구현체 분리
